feat: member-product-inquiry-view
