+++ http/member-test.http
@@ -4,8 +4,8 @@ Content-Type: application/json
 
 {
 
-  "memberId" : "user03",
-  "memberPassword" : "pass03",
+  "memberId" : "user02",
+  "memberPassword" : "pass02",
   "memberName": "회원",
   "memberEmail": "user02@greenfire.com",
   "memberPhone": "01012341234"
@@ -16,8 +16,8 @@ POST http://localhost:8001/members/login
 Content-Type: application/json
 
 {
-  "memberId" : "user01",
-  "memberPassword" : "pass01"
+  "memberId" : "user02",
+  "memberPassword" : "pass02"
 }
 
 ### 인증 되어야 하는 기능을 대상으로 테스트
