+++ src/main/java/yep/greenFire/greenfirebackend/inquiry/entity/InquiryContent.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import yep.greenFire.greenfirebackend.challenge.dto.response.inquiry.AdminInquiryResponse;
+import yep.greenFire.greenfirebackend.inquiry.site.admin.dto.response.AdminInquiryResponse;
 
 import java.util.Date;
 
@@ -30,8 +30,14 @@ public class InquiryContent {
     @LastModifiedDate
     private Date inquiryModifyDate;
     private Date inquiryDeleteDate;
+    private int replyMemberCode;
     private String inquiryReply;
     private String inquiryReplyStatus;
+
+    private int inquiryCateCode;
+
+    private int inquiryProduct;
+    private int inquiryChallenge;
