+++ src/main/java/yep/greenFire/greenfirebackend/inquiry/product/member/domain/repository/MemberProductInquiryRepository.java
@@ -1,0 +1,14 @@
+package yep.greenFire.greenfirebackend.inquiry.product.member.domain.repository;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Repository;
+import yep.greenFire.greenfirebackend.inquiry.entity.InquiryContent;
+
+
+@Repository
+public interface MemberProductInquiryRepository {
+
+    Page<InquiryContent> findByProductCode(long productCode, Pageable pageable);
+
+}
